Poured ice into the input blender

+++ TwelveEngine/Input/Impulse.cs
@@ -1,5 +1,5 @@
 ﻿namespace TwelveEngine {
-    public enum Impulse:byte {
+    public enum Impulse:int {
         /* Don't change the order of the key binds enum.
          * A strict order allows for backwards compatibility */
         Up, Down, Left, Right, Accept, Exit
